Make the update strategy not Immix-specific
