+++ src/vm/mod.rs
@@ -27,8 +27,6 @@ pub use self::active_plan::ActivePlan;
 pub use self::collection::Collection;
 pub use self::collection::GCThreadContext;
 pub use self::object_model::specs::*;
-#[cfg(feature = "vo_bit")]
-pub use self::object_model::ImmixVOBitUpdateStrategy;
 pub use self::object_model::ObjectModel;
 pub use self::reference_glue::Finalizable;
 pub use self::reference_glue::ReferenceGlue;
